+++ src/main/kotlin/com/github/saundefined/bitrix_idea/actions/CreateSimpleComponentAction.kt
@@ -1,10 +1,13 @@
 package com.github.saundefined.bitrix_idea.actions
 
 import com.github.saundefined.bitrix_idea.BitrixIdeaBundle.message
+import com.github.saundefined.bitrix_idea.dialogs.CreateSimpleComponentDialogWrapper
+import com.intellij.ide.util.DirectoryChooserUtil
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
-import com.intellij.psi.PsiDirectory
+import com.intellij.openapi.actionSystem.LangDataKeys
+import com.intellij.openapi.project.Project
 import icons.license.CheckLicense
 import org.jetbrains.annotations.NotNull
 
@@ -15,6 +18,27 @@ class CreateSimpleComponentAction : AnAction() {
             CheckLicense.requestLicense(message("license.request"))
             return
         }
+
+        val view = event.getData(LangDataKeys.IDE_VIEW)
+        if (view === null) {
+            return
+        }
+
+        val project: Project? = event.project
+        if (project === null) {
+            return
+        }
+
+        val directory = DirectoryChooserUtil.getOrChooseDirectory(view)
+        if (directory === null) {
+            return
+        }
+
+        val dialogWrapper = CreateSimpleComponentDialogWrapper()
+        dialogWrapper.view = view
+        dialogWrapper.project = project
+        dialogWrapper.directory = directory
+        dialogWrapper.showAndGet()
     }
 
     override fun getActionUpdateThread(): ActionUpdateThread {